+++ src/domain/eventGroups/edit/EventGroupsEdit.tsx
@@ -1,7 +1,15 @@
 import React from 'react';
+import { ResourceComponentProps } from 'react-admin';
 
-const EventGroupsEdit = () => {
-  return <>EventGroupsEdit</>;
+import KukkuuEditPage from '../../application/layout/kukkuuEditPage/KukkuuEditPage';
+import EventGroupForm from '../form/EventGroupForm';
+
+const EventGroupsEdit = (props: ResourceComponentProps) => {
+  return (
+    <KukkuuEditPage pageTitleSource="name" reactAdminProps={props}>
+      <EventGroupForm redirect="/events-and-event-groups" />
+    </KukkuuEditPage>
+  );
 };
 
 export default EventGroupsEdit;